+++ teku/src/test/java/tech/pegasys/teku/cli/options/Eth2NetworkOptionsTest.java
@@ -33,11 +33,11 @@ import tech.pegasys.teku.spec.config.ProgressiveBalancesMode;
 
 class Eth2NetworkOptionsTest extends AbstractBeaconNodeCommandTest {
   @Test
-  void shouldEnableBellatrixByDefault() {
+  void shouldEnableCapellaByDefault() {
     final TekuConfiguration config = getTekuConfigurationFromArguments();
     final Spec spec = config.eth2NetworkConfiguration().getSpec();
     assertThat(spec.getForkSchedule().getHighestSupportedMilestone())
-        .isEqualTo(SpecMilestone.BELLATRIX);
+        .isEqualTo(SpecMilestone.CAPELLA);
   }
 
   @Test
@@ -97,11 +97,22 @@ class Eth2NetworkOptionsTest extends AbstractBeaconNodeCommandTest {
 
   @ParameterizedTest
   @EnumSource(value = ProgressiveBalancesMode.class)
-  void shouldUseCustomProgressiveBalancesMode(final ProgressiveBalancesMode mode) {
+  void shouldUseCustomProgressiveBalancesModeOnMinimal(final ProgressiveBalancesMode mode) {
+    final TekuConfiguration config =
+        getTekuConfigurationFromArguments(
+            "--network", "minimal", "--Xprogressive-balances-mode", mode.name());
+    final Spec spec = config.eth2NetworkConfiguration().getSpec();
+    assertThat(spec.getGenesisSpecConfig().getProgressiveBalancesMode()).isEqualTo(mode);
+  }
+
+  @ParameterizedTest
+  @EnumSource(value = ProgressiveBalancesMode.class)
+  void shouldNotUseCustomProgressiveBalancesModeOnMainnet(final ProgressiveBalancesMode mode) {
     final TekuConfiguration config =
         getTekuConfigurationFromArguments("--Xprogressive-balances-mode", mode.name());
     final Spec spec = config.eth2NetworkConfiguration().getSpec();
-    assertThat(spec.getGenesisSpecConfig().getProgressiveBalancesMode()).isEqualTo(mode);
+    assertThat(spec.getGenesisSpecConfig().getProgressiveBalancesMode())
+        .isEqualTo(ProgressiveBalancesMode.FULL);
   }
 
   @ParameterizedTest